chatbot/requirements.txt
fastapi>=0.110.0
uvicorn>=0.24.0

python-dotenv>=1.0.0

qdrant-client>=1.10.1

openai>=1.30.0
cohere>=4.37.0

tiktoken>=0.5.2
markdown>=3.5.1
beautifulsoup4>=4.12.2

pydantic>=2.6.0

requests>=2.31.0
